Move Host models to Contract
