+++ Auth.sln
@@ -8,6 +8,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelliCook.Auth.Host.E2ETe
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelliCook.Auth.Host.UnitTests", "IntelliCook.Auth.Host.UnitTests\IntelliCook.Auth.Host.UnitTests.csproj", "{AB97F45A-796A-4027-BB88-7D674311CF09}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelliCook.Auth.Contract", "IntelliCook.Auth.Contract\IntelliCook.Auth.Contract.csproj", "{670E355B-5DD6-44C9-B89A-14E39CCE235A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +32,9 @@ Global
 		{AB97F45A-796A-4027-BB88-7D674311CF09}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AB97F45A-796A-4027-BB88-7D674311CF09}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AB97F45A-796A-4027-BB88-7D674311CF09}.Release|Any CPU.Build.0 = Release|Any CPU
+		{670E355B-5DD6-44C9-B89A-14E39CCE235A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{670E355B-5DD6-44C9-B89A-14E39CCE235A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{670E355B-5DD6-44C9-B89A-14E39CCE235A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{670E355B-5DD6-44C9-B89A-14E39CCE235A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
