+++ IntelliCook.Auth.Contract/Health/HealthStatusModel.cs
@@ -1,0 +1,8 @@
+namespace IntelliCook.Auth.Contract.Health;
+
+public enum HealthStatusModel
+{
+    Healthy,
+    Degraded,
+    Unhealthy
+}
