+++ IntelliCook.Auth.Host.E2ETests/Fixtures/ClientFixture.cs
@@ -1,4 +1,4 @@
-using IntelliCook.Auth.Host.Models.Auth.Register;
+using IntelliCook.Auth.Contract.Auth.Register;
 using IntelliCook.Auth.Host.Options;
 using IntelliCook.Auth.Infrastructure.Models;
 using Microsoft.AspNetCore.Mvc.Testing;
