+++ IntelliCook.Auth.Host/Controllers/Auth/LoginController.cs
@@ -1,4 +1,4 @@
-using IntelliCook.Auth.Host.Models.Auth.Login;
+using IntelliCook.Auth.Contract.Auth.Login;
 using IntelliCook.Auth.Host.Options;
 using IntelliCook.Auth.Infrastructure.Models;
 using Microsoft.AspNetCore.Authorization;
